post working properly with delete and update

+++ app/Http/Controllers/AdminUsersController.php
@@ -144,8 +144,9 @@ class AdminUsersController extends Controller
         //
         $user = User::findOrFail($id);
 
-        unlink(public_path(). $user->photo->file);
-
+        if($user->photo){
+            unlink(public_path(). $user->photo->file);
+        }
         $user->delete();
         Session::flash('deleted_user','The user has been deleted');
         return redirect('/admin/users');
